* develop:
  Fix listing issues
  Add --verbose & --debug flags
  Add logging mechanism
  Add utils
  Add files
  Switch to using Thor for CLI
  Add inch documentation badge
  Add dependencies badge to README

+++ lib/dotstrap/shell.rb
@@ -1,3 +1,5 @@
+require 'logger'
+
 module Dotstrap
   class Dotstrap::Shell
     # TODO: cleanup shell class & split it into 3+ classes with inheritence
@@ -9,25 +11,28 @@ module Dotstrap
     end
 
     def configure(repo_dir = @repo_path)
-      configure_fish(repo_dir) if Dir.exist?(fish_config_home)
+      configure_fish(repo_dir) if Dir.exist?(Dotstrap.shell_profile('fish'))
       configure_zsh(repo_dir)
       configure_bash(repo_dir)
+      $LOG.unknown { "configuration complete" }
     end
 
     def unconfigure(repo_dir = @repo_path)
-      unconfigure_fish(repo_dir) if Dir.exist?(fish_config_home)
+      return unless Dir.exist?(repo_dir)
+      unconfigure_fish(repo_dir) if Dir.exist?(Dotstrap.shell_profile('fish'))
       unconfigure_zsh(repo_dir)
       unconfigure_bash(repo_dir)
       FileUtils.rm_r(repo_dir, force: true, secure: true)
+      $LOG.unknown { "removed: #{repo_dir}\n" }
     end
 
     def configure_fish(repo_dir)
       fish_functions(repo_dir).each do |f|
-        link_config_file(f, File.join(fish_config_home, 'functions'))
+        link_config_file(f, File.join(Dotstrap.shell_profile('fish'), 'functions'))
       end
 
       fish_completions(repo_dir).each do |f|
-        link_config_file(f, File.join(fish_config_home, 'completions'))
+        link_config_file(f, File.join(Dotstrap.shell_profile('fish'), 'completions'))
       end
 
       write_config_file(fish_configs(repo_dir), config_file('fish'))
@@ -35,12 +40,12 @@ module Dotstrap
 
     def unconfigure_fish(repo_dir)
       fish_functions(repo_dir).each do |f|
-        file = File.join(fish_config_home, 'functions', File.basename(f))
+        file = File.join(Dotstrap.shell_profile('fish'), 'functions', File.basename(f))
         rm_config_file(file)
       end
 
       fish_completions(repo_dir).each do |f|
-        file = File.join(fish_config_home, 'completions', File.basename(f))
+        file = File.join(Dotstrap.shell_profile('fish'), 'completions', File.basename(f))
         rm_config_file(file)
       end
 
@@ -69,7 +74,7 @@ module Dotstrap
       FileUtils.rm_rf file
     end
 
-    def config_file(sh, dir = Dotstrap.dotstrap_config_home)
+    def config_file(sh, dir = Dotstrap.config_home)
       File.join(dir, "config.#{sh}")
     end
 
@@ -79,6 +84,7 @@ module Dotstrap
       unless repo_config_files.empty?
         repo_config_files.each do |file|
           append_config_file(file, dotstrap_config_file)
+          $LOG.info { "configured: #{file}" }
         end
       end
     end
@@ -87,6 +93,7 @@ module Dotstrap
       unless repo_config_files.empty?
         repo_config_files.each do |file|
           unappend_config_file(file, dotstrap_config_file)
+          $LOG.info { "removed: #{file}" }
         end
       end
     end
@@ -107,7 +114,7 @@ module Dotstrap
         source_str = "source \"#{file_to_unsource}\""
         tmp_config_file = File.join(tmp_dir, File.basename(config_file))
         File.open(tmp_config_file, 'w') do |tmp_file|
-          File.foreach(input_file) do |line|
+          File.foreach(file_to_unsource) do |line|
             tmp_file.write line unless line.chomp == source_str
           end
         end
@@ -122,11 +129,7 @@ module Dotstrap
       # FileUtils.mkdir parent unless Dir.exist? parent
       FileUtils.mkdir_p dest_dir unless Dir.exist? dest_dir
       FileUtils.ln_s src, dst, force: true
-    end
-
-    def fish_config_home
-      config_home = ENV.fetch('XDG_CONFIG_HOME', File.expand_path('~/.config'))
-      File.join(config_home, 'fish')
+      $LOG.info { "linked: #{src} to #{dst}" }
     end
 
     def fish_configs(dir)
